website/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.2.0
defusedxml==0.7.1
Django==4.2.3
django-cors-headers==4.1.0
django-extensions==3.2.3
django-filter==23.2
django-leaflet==0.29.0
djangorestframework==3.14.0
djangorestframework-gis==1.0
djangorestframework-simplejwt==5.2.2
gunicorn
holidays==0.29
httpie==3.2.2
idna==3.4
joblib==1.3.1
Markdown==3.4.3
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.4
numpy==1.25.1
pandas==2.0.3
pgeocode==0.4.0
psycopg2-binary==2.9.6
pyarrow==12.0.1
Pygments==2.15.1
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.4.2
scikit-learn
scipy==1.11.1
six==1.16.0
scikit-learn
sodapy==2.2.0
sqlparse==0.4.4
threadpoolctl==3.2.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
xgboost==1.7.6
